+++ uppoint-cloud/app/[locale]/dashboard/security/page.tsx
@@ -19,8 +19,8 @@ export async function generateMetadata({
   const { metadata } = getDictionary(locale);
 
   return {
-    title: metadata.dashboard.title,
-    description: metadata.dashboard.description,
+    title: metadata.dashboardSecurity.title,
+    description: metadata.dashboardSecurity.description,
   };
 }